delete config.py
