+++ project/server/__init__.py
@@ -1,6 +1,6 @@
 """ This script will convert the api to a model """
 from flask import Flask
-from config import DevelopmentConfig, TestingConfig, ProductionConfig
+from project.server.config import DevelopmentConfig, TestingConfig, ProductionConfig
 
 app = Flask(__name__)
 app.config.from_object(TestingConfig)
